Implement registering a fantasy team
Implement commands and database interactions for registering a fantasy team. WIP: NOT TESTED. Make team and abbrev fields in fantasy_teams table unique. Use fetchone instead of fetchall for queries with expected result of a single row. Place onus of committing a transaction on the command, rather than the query.