Fantasy database tables and example command (#9)
* Fantasy team tables

Write creation statements for fantasy team tables. Not yet reflected in actual database.

* Make certain rows unique

Make certain rows unique.

* Mod to users table

Make primary key discord_id.

* Add example fantasy queries

Add a few example fantasy queries. Also serve as examples for insert and update queries.

* Implement registering a fantasy team

Implement commands and database interactions for registering a fantasy team. WIP: NOT TESTED. Make team and abbrev fields in fantasy_teams table unique. Use fetchone instead of fetchall for queries with expected result of a single row. Place onus of committing a transaction on the command, rather than the query.

* fix typo

* bugfix: forgot import

* bugfix: bot can't talk

* !register handle existing user

Make allowance if the user is already in the database, but does not have
a team assigned to them.

* bugfix: typo

* bugfix: typo

* bugfix: typo

* bugfix: ensure object won't be null